+++ packages/applications/document-builder/src/candidature/attestation/attestation.saga.ts
@@ -100,8 +100,6 @@ export const register = () => {
           format,
         );
 
-        console.log('violette build', attestation);
-
         await mediator.send<EnregistrerDocumentProjetCommand>({
           type: 'Document.Command.EnregistrerDocumentProjet',
           data: {
