+++ packages/applications/legacy/src/controllers/notificationCandidats/index.ts
@@ -1,5 +1,4 @@
 export * from './legacyGetAdminImporterCandidatsPage';
 export * from './getAdminNotificationCandidatsPage';
-export * from './getProjectCertificateFile';
 export * from './postImportProjects';
 export * from './postSendCandidateNotifications';
